Move the exceptions loading code into util
