+++ salesforce2OneCRL/salesforce2OneCRL.go
@@ -9,79 +9,20 @@ import (
 	"encoding/asn1"
 	"encoding/base64"
 	"encoding/hex"
-	"encoding/json"
 	"encoding/pem"
-	"errors"
 	"flag"
 	"fmt"
 	constraintsx509 "github.com/jcjones/constraintcrypto/x509"
 	"github.com/mozilla/OneCRL-Tools/config"
 	"github.com/mozilla/OneCRL-Tools/oneCRL"
 	"github.com/mozilla/OneCRL-Tools/salesforce"
+	"github.com/mozilla/OneCRL-Tools/util"
 	"io"
-	"io/ioutil"
 	"net/http"
 	"os"
 	"strings"
 	"time"
 )
-
-func getDataFromURL(url string) ([]byte, error) {
-	r, _ := http.Get(url)
-	defer r.Body.Close()
-
-	return ioutil.ReadAll(r.Body)
-}
-
-func recordExists(item oneCRL.Record, records *oneCRL.Records) bool {
-	for _, record := range records.Data {
-		if item.EqualsRecord(record) {
-			return true
-		}
-	}
-	return false
-}
-
-func LoadExceptions(location string, existing *oneCRL.Records, records *oneCRL.Records) error {
-	res := new(oneCRL.Records)
-	var data []byte
-
-	if 0 != strings.Index(strings.ToUpper(location), "HTTP") {
-		// if it's not an HTTP URL, attempt to load from a file
-		if fileData, err := ioutil.ReadFile(location); nil != err {
-			fmt.Printf("problem loading oneCRL exceptions from file %s\n", err)
-		} else {
-			data = fileData
-		}
-	} else {
-		// ensure it's not an HTTP location
-		if 0 != strings.Index(strings.ToUpper(location), "HTTPS") {
-			return errors.New("Cowardly refusing to load exceptions from a non HTTPS location")
-		}
-		if resp, err := http.Get(location); nil != err {
-			return err
-		} else {
-			defer resp.Body.Close()
-			if urlData, err := ioutil.ReadAll(resp.Body); nil != err {
-				return err
-			} else {
-				data = urlData
-			}
-		}
-	}
-
-	if err := json.Unmarshal(data, res); nil != err {
-		return err
-	}
-
-	for idx := range res.Data {
-		record := res.Data[idx]
-		if !recordExists(record, existing) {
-			records.Data = append(records.Data, record)
-		}
-	}
-	return nil
-}
 
 func main() {
 	filePtr := flag.String("file", "", "The file to read data from")
@@ -134,7 +75,7 @@ func main() {
 	}
 
 	if len(*exceptionsPtr) != 0 {
-		if err := LoadExceptions(*exceptionsPtr, existing, additions); nil != err {
+		if err := util.LoadExceptions(*exceptionsPtr, existing, additions); nil != err {
 			panic(err)
 		}
 	}
@@ -168,7 +109,7 @@ func main() {
 			serialString := base64.StdEncoding.EncodeToString(serialBytes[2:])
 
 			record := oneCRL.Record{IssuerName: issuerString, SerialNumber: serialString}
-			if recordExists(record, existing) {
+			if util.RecordExists(record, existing) {
 				fmt.Printf("(%d, %s, %s) revocation already in OneCRL\n", row, each.CSN, each.CertName)
 				continue
 			}
